backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<4.2.0
pymysql>=1.1.0
cryptography>=42.0.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
alembic>=1.14.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
httpx>=0.27.0
email-validator>=2.0.0
